InstallerController now take the user through the entire installation process, and tries to give informative error messages. it also present the user with a welcome page when setup are done.
app.php is now updated with database connection details during setup.

SetupSimplicityForm is a 'table-less' form with field validation rules and the whole fadderull.

cleanup in Layout/installer.ctp, it now works with Zurb Foundation and with jQuery.

each Table class now has function CreateTable() which are creating it's table in the database.