sidemenu visible and working for logged in users, currently showing the administrator tools.
MenuComponent::CreateMenuElement() is useful when creating menu-structures when the content is _not_ RichTextElements.

MenuComponent now create full url to element. (So MenuHelper don't have to)

MenuHelper::GetAccordionMenu() now has a small hack so the menu can open up, and then transform itself to a clickable link.

css-fix for top-bar, it needed flex-shrink: 0; as well to make the footer-oriented design work.