title and url_title are now used instead of old 'name', giving the possibility to have a nicer-looking title in the menus, while the url_title can stay ugly and/or non-translated.
RichTextElementsTable.page_id keeps the different translated versions of a page together.

if the home-page is deleted (or never created) it is faked, instead of giving ugly errors.

admin can now edit url_title and title of a RichTextElement. Exception is home, where the system informs that url_title cannot be messed with.